script should fail if soot fails

+++ bin/scripts/microbenchmark/driver_generator_microbenchmark.sh
@@ -32,7 +32,11 @@ if [ -a is_maven ]; then
   mvn clean test
 fi
 
-java -cp $SOOT_JAR:$JAVA_PATH $CC_CLASS $BENCHMARK_PATH/$TEXT_PATH
+echo java -cp $SOOT_JAR:$JAVA_PATH $CC_CLASS $BENCHMARK_PATH/$TEXT_PATH
+if ! java -cp $SOOT_JAR:$JAVA_PATH $CC_CLASS $BENCHMARK_PATH/$TEXT_PATH; then
+  echo "soot failed"
+  exit 1
+fi
 rm -rf "sootOutput/"
 
-mvn package
+mvn -Drat.skip=true package
